Add user interactors
